Updated lens dialogs and changes how db access worked
