+++ app/src/main/res/layout/lenslist_layout.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+    <ScrollView
+        android:id="@+id/scrollView7"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_gravity="center_horizontal"
+        android:gravity="center"
+        android:layoutMode="clipBounds">
+
+        <GridLayout
+            android:id="@+id/lensloader_gridholder"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center_horizontal"
+            android:alignmentMode="alignBounds"
+            android:columnCount="4"
+            android:layoutMode="clipBounds"
+            android:orientation="horizontal"
+            android:useDefaultMargins="false"></GridLayout>
+    </ScrollView>
+</LinearLayout>
